dataset/python-mutated/theanocode.py
"""
.. deprecated:: 1.8

  ``sympy.printing.theanocode`` is deprecated. Theano has been renamed to
  Aesara. Use ``sympy.printing.aesaracode`` instead. See
  :ref:`theanocode-deprecated` for more information.

"""
from __future__ import annotations
from typing import Any
from sympy.external import import_module
from sympy.printing.printer import Printer
from sympy.utilities.iterables import is_sequence
import sympy
from functools import partial
from sympy.utilities.decorator import doctest_depends_on
from sympy.utilities.exceptions import sympy_deprecation_warning
theano = import_module('theano')
if theano:
    ts = theano.scalar
    tt = theano.tensor
    from theano.sandbox import linalg as tlinalg
    mapping = {sympy.Add: tt.add, sympy.Mul: tt.mul, sympy.Abs: tt.abs_, sympy.sign: tt.sgn, sympy.ceiling: tt.ceil, sympy.floor: tt.floor, sympy.log: tt.log, sympy.exp: tt.exp, sympy.sqrt: tt.sqrt, sympy.cos: tt.cos, sympy.acos: tt.arccos, sympy.sin: tt.sin, sympy.asin: tt.arcsin, sympy.tan: tt.tan, sympy.atan: tt.arctan, sympy.atan2: tt.arctan2, sympy.cosh: tt.cosh, sympy.acosh: tt.arccosh, sympy.sinh: tt.sinh, sympy.asinh: tt.arcsinh, sympy.tanh: tt.tanh, sympy.atanh: tt.arctanh, sympy.re: tt.real, sympy.im: tt.imag, sympy.arg: tt.angle, sympy.erf: tt.erf, sympy.gamma: tt.gamma, sympy.loggamma: tt.gammaln, sympy.Pow: tt.pow, sympy.Eq: tt.eq, sympy.StrictGreaterThan: tt.gt, sympy.StrictLessThan: tt.lt, sympy.LessThan: tt.le, sympy.GreaterThan: tt.ge, sympy.And: tt.and_, sympy.Or: tt.or_, sympy.Max: tt.maximum, sympy.Min: tt.minimum, sympy.conjugate: tt.conj, sympy.core.numbers.ImaginaryUnit: lambda : tt.complex(0, 1), sympy.MatAdd: tt.Elemwise(ts.add), sympy.HadamardProduct: tt.Elemwise(ts.mul), sympy.Trace: tlinalg.trace, sympy.Determinant: tlinalg.det, sympy.Inverse: tlinalg.matrix_inverse, sympy.Transpose: tt.DimShuffle((False, False), [1, 0])}

class TheanoPrinter(Printer):
    """ Code printer which creates Theano symbolic expression graphs.

    Parameters
    ==========

    cache : dict
        Cache dictionary to use. If None (default) will use
        the global cache. To create a printer which does not depend on or alter
        global state pass an empty dictionary. Note: the dictionary is not
        copied on initialization of the printer and will be updated in-place,
        so using the same dict object when creating multiple printers or making
        multiple calls to :func:`.theano_code` or :func:`.theano_function` means
        the cache is shared between all these applications.

    Attributes
    ==========

    cache : dict
        A cache of Theano variables which have been created for SymPy
        symbol-like objects (e.g. :class:`sympy.core.symbol.Symbol` or
        :class:`sympy.matrices.expressions.MatrixSymbol`). This is used to
        ensure that all references to a given symbol in an expression (or
        multiple expressions) are printed as the same Theano variable, which is
        created only once. Symbols are differentiated only by name and type. The
        format of the cache's contents should be considered opaque to the user.
    """
    printmethod = '_theano'

    def __init__(self, *args, **kwargs):
        if False:
            return 10
        self.cache = kwargs.pop('cache', {})
        super().__init__(*args, **kwargs)

    def _get_key(self, s, name=None, dtype=None, broadcastable=None):
        if False:
            while True:
                i = 10
        ' Get the cache key for a SymPy object.\n\n        Parameters\n        ==========\n\n        s : sympy.core.basic.Basic\n            SymPy object to get key for.\n\n        name : str\n            Name of object, if it does not have a ``name`` attribute.\n        '
        if name is None:
            name = s.name
        return (name, type(s), s.args, dtype, broadcastable)

    def _get_or_create(self, s, name=None, dtype=None, broadcastable=None):
        if False:
            for i in range(10):
                print('nop')
        '\n        Get the Theano variable for a SymPy symbol from the cache, or create it\n        if it does not exist.\n        '
        if name is None:
            name = s.name
        if dtype is None:
            dtype = 'floatX'
        if broadcastable is None:
            broadcastable = ()
        key = self._get_key(s, name, dtype=dtype, broadcastable=broadcastable)
        if key in self.cache:
            return self.cache[key]
        value = tt.tensor(name=name, dtype=dtype, broadcastable=broadcastable)
        self.cache[key] = value
        return value

    def _print_Symbol(self, s, **kwargs):
        if False:
            print('Hello World!')
        dtype = kwargs.get('dtypes', {}).get(s)
        bc = kwargs.get('broadcastables', {}).get(s)
        return self._get_or_create(s, dtype=dtype, broadcastable=bc)

    def _print_AppliedUndef(self, s, **kwargs):
        if False:
            print('Hello World!')
        name = str(type(s)) + '_' + str(s.args[0])
        dtype = kwargs.get('dtypes', {}).get(s)
        bc = kwargs.get('broadcastables', {}).get(s)
        return self._get_or_create(s, name=name, dtype=dtype, broadcastable=bc)

    def _print_Basic(self, expr, **kwargs):
        if False:
            for i in range(10):
                print('nop')
        op = mapping[type(expr)]
        children = [self._print(arg, **kwargs) for arg in expr.args]
        return op(*children)

    def _print_Number(self, n, **kwargs):
        if False:
            while True:
                i = 10
        return float(n.evalf())

    def _print_MatrixSymbol(self, X, **kwargs):
        if False:
            return 10
        dtype = kwargs.get('dtypes', {}).get(X)
        return self._get_or_create(X, dtype=dtype, broadcastable=(None, None))

    def _print_DenseMatrix(self, X, **kwargs):
        if False:
            for i in range(10):
                print('nop')
        if not hasattr(tt, 'stacklists'):
            raise NotImplementedError('Matrix translation not yet supported in this version of Theano')
        return tt.stacklists([[self._print(arg, **kwargs) for arg in L] for L in X.tolist()])
    _print_ImmutableMatrix = _print_ImmutableDenseMatrix = _print_DenseMatrix

    def _print_MatMul(self, expr, **kwargs):
        if False:
            print('Hello World!')
        children = [self._print(arg, **kwargs) for arg in expr.args]
        result = children[0]
        for child in children[1:]:
            result = tt.dot(result, child)
        return result

    def _print_MatPow(self, expr, **kwargs):
        if False:
            for i in range(10):
                print('nop')
        children = [self._print(arg, **kwargs) for arg in expr.args]
        result = 1
        if isinstance(children[1], int) and children[1] > 0:
            for i in range(children[1]):
                result = tt.dot(result, children[0])
        else:
            raise NotImplementedError('Only non-negative integer\n           powers of matrices can be handled by Theano at the moment')
        return result

    def _print_MatrixSlice(self, expr, **kwargs):
        if False:
            return 10
        parent = self._print(expr.parent, **kwargs)
        rowslice = self._print(slice(*expr.rowslice), **kwargs)
        colslice = self._print(slice(*expr.colslice), **kwargs)
        return parent[rowslice, colslice]

    def _print_BlockMatrix(self, expr, **kwargs):
        if False:
            for i in range(10):
                print('nop')
        (nrows, ncols) = expr.blocks.shape
        blocks = [[self._print(expr.blocks[r, c], **kwargs) for c in range(ncols)] for r in range(nrows)]
        return tt.join(0, *[tt.join(1, *row) for row in blocks])

    def _print_slice(self, expr, **kwargs):
        if False:
            print('Hello World!')
        return slice(*[self._print(i, **kwargs) if isinstance(i, sympy.Basic) else i for i in (expr.start, expr.stop, expr.step)])

    def _print_Pi(self, expr, **kwargs):
        if False:
            for i in range(10):
                print('nop')
        return 3.141592653589793

    def _print_Exp1(self, expr, **kwargs):
        if False:
            for i in range(10):
                print('nop')
        return ts.exp(1)

    def _print_Piecewise(self, expr, **kwargs):
        if False:
            print('Hello World!')
        import numpy as np
        (e, cond) = expr.args[0].args
        p_cond = self._print(cond, **kwargs)
        p_e = self._print(e, **kwargs)
        if len(expr.args) == 1:
            return tt.switch(p_cond, p_e, np.nan)
        p_remaining = self._print(sympy.Piecewise(*expr.args[1:]), **kwargs)
        return tt.switch(p_cond, p_e, p_remaining)

    def _print_Rational(self, expr, **kwargs):
        if False:
            print('Hello World!')
        return tt.true_div(self._print(expr.p, **kwargs), self._print(expr.q, **kwargs))

    def _print_Integer(self, expr, **kwargs):
        if False:
            while True:
                i = 10
        return expr.p

    def _print_factorial(self, expr, **kwargs):
        if False:
            print('Hello World!')
        return self._print(sympy.gamma(expr.args[0] + 1), **kwargs)

    def _print_Derivative(self, deriv, **kwargs):
        if False:
            i = 10
            return i + 15
        rv = self._print(deriv.expr, **kwargs)
        for var in deriv.variables:
            var = self._print(var, **kwargs)
            rv = tt.Rop(rv, var, tt.ones_like(var))
        return rv

    def emptyPrinter(self, expr):
        if False:
            i = 10
            return i + 15
        return expr

    def doprint(self, expr, dtypes=None, broadcastables=None):
        if False:
            i = 10
            return i + 15
        " Convert a SymPy expression to a Theano graph variable.\n\n        The ``dtypes`` and ``broadcastables`` arguments are used to specify the\n        data type, dimension, and broadcasting behavior of the Theano variables\n        corresponding to the free symbols in ``expr``. Each is a mapping from\n        SymPy symbols to the value of the corresponding argument to\n        ``theano.tensor.Tensor``.\n\n        See the corresponding `documentation page`__ for more information on\n        broadcasting in Theano.\n\n        .. __: http://deeplearning.net/software/theano/tutorial/broadcasting.html\n\n        Parameters\n        ==========\n\n        expr : sympy.core.expr.Expr\n            SymPy expression to print.\n\n        dtypes : dict\n            Mapping from SymPy symbols to Theano datatypes to use when creating\n            new Theano variables for those symbols. Corresponds to the ``dtype``\n            argument to ``theano.tensor.Tensor``. Defaults to ``'floatX'``\n            for symbols not included in the mapping.\n\n        broadcastables : dict\n            Mapping from SymPy symbols to the value of the ``broadcastable``\n            argument to ``theano.tensor.Tensor`` to use when creating Theano\n            variables for those symbols. Defaults to the empty tuple for symbols\n            not included in the mapping (resulting in a scalar).\n\n        Returns\n        =======\n\n        theano.gof.graph.Variable\n            A variable corresponding to the expression's value in a Theano\n            symbolic expression graph.\n\n        "
        if dtypes is None:
            dtypes = {}
        if broadcastables is None:
            broadcastables = {}
        return self._print(expr, dtypes=dtypes, broadcastables=broadcastables)
global_cache: dict[Any, Any] = {}

def theano_code(expr, cache=None, **kwargs):
    if False:
        for i in range(10):
            print('nop')
    "\n    Convert a SymPy expression into a Theano graph variable.\n\n    .. deprecated:: 1.8\n\n      ``sympy.printing.theanocode`` is deprecated. Theano has been renamed to\n      Aesara. Use ``sympy.printing.aesaracode`` instead. See\n      :ref:`theanocode-deprecated` for more information.\n\n    Parameters\n    ==========\n\n    expr : sympy.core.expr.Expr\n        SymPy expression object to convert.\n\n    cache : dict\n        Cached Theano variables (see :class:`TheanoPrinter.cache\n        <TheanoPrinter>`). Defaults to the module-level global cache.\n\n    dtypes : dict\n        Passed to :meth:`.TheanoPrinter.doprint`.\n\n    broadcastables : dict\n        Passed to :meth:`.TheanoPrinter.doprint`.\n\n    Returns\n    =======\n\n    theano.gof.graph.Variable\n        A variable corresponding to the expression's value in a Theano symbolic\n        expression graph.\n\n    "
    sympy_deprecation_warning('\n        sympy.printing.theanocode is deprecated. Theano has been renamed to\n        Aesara. Use sympy.printing.aesaracode instead.', deprecated_since_version='1.8', active_deprecations_target='theanocode-deprecated')
    if not theano:
        raise ImportError('theano is required for theano_code')
    if cache is None:
        cache = global_cache
    return TheanoPrinter(cache=cache, settings={}).doprint(expr, **kwargs)

def dim_handling(inputs, dim=None, dims=None, broadcastables=None):
    if False:
        i = 10
        return i + 15
    '\n    Get value of ``broadcastables`` argument to :func:`.theano_code` from\n    keyword arguments to :func:`.theano_function`.\n\n    Included for backwards compatibility.\n\n    Parameters\n    ==========\n\n    inputs\n        Sequence of input symbols.\n\n    dim : int\n        Common number of dimensions for all inputs. Overrides other arguments\n        if given.\n\n    dims : dict\n        Mapping from input symbols to number of dimensions. Overrides\n        ``broadcastables`` argument if given.\n\n    broadcastables : dict\n        Explicit value of ``broadcastables`` argument to\n        :meth:`.TheanoPrinter.doprint`. If not None function will return this value unchanged.\n\n    Returns\n    =======\n    dict\n        Dictionary mapping elements of ``inputs`` to their "broadcastable"\n        values (tuple of ``bool``\\ s).\n    '
    if dim is not None:
        return {s: (False,) * dim for s in inputs}
    if dims is not None:
        maxdim = max(dims.values())
        return {s: (False,) * d + (True,) * (maxdim - d) for (s, d) in dims.items()}
    if broadcastables is not None:
        return broadcastables
    return {}

@doctest_depends_on(modules=('theano',))
def theano_function(inputs, outputs, scalar=False, *, dim=None, dims=None, broadcastables=None, **kwargs):
    if False:
        print('Hello World!')
    '\n    Create a Theano function from SymPy expressions.\n\n    .. deprecated:: 1.8\n\n      ``sympy.printing.theanocode`` is deprecated. Theano has been renamed to\n      Aesara. Use ``sympy.printing.aesaracode`` instead. See\n      :ref:`theanocode-deprecated` for more information.\n\n    The inputs and outputs are converted to Theano variables using\n    :func:`.theano_code` and then passed to ``theano.function``.\n\n    Parameters\n    ==========\n\n    inputs\n        Sequence of symbols which constitute the inputs of the function.\n\n    outputs\n        Sequence of expressions which constitute the outputs(s) of the\n        function. The free symbols of each expression must be a subset of\n        ``inputs``.\n\n    scalar : bool\n        Convert 0-dimensional arrays in output to scalars. This will return a\n        Python wrapper function around the Theano function object.\n\n    cache : dict\n        Cached Theano variables (see :class:`TheanoPrinter.cache\n        <TheanoPrinter>`). Defaults to the module-level global cache.\n\n    dtypes : dict\n        Passed to :meth:`.TheanoPrinter.doprint`.\n\n    broadcastables : dict\n        Passed to :meth:`.TheanoPrinter.doprint`.\n\n    dims : dict\n        Alternative to ``broadcastables`` argument. Mapping from elements of\n        ``inputs`` to integers indicating the dimension of their associated\n        arrays/tensors. Overrides ``broadcastables`` argument if given.\n\n    dim : int\n        Another alternative to the ``broadcastables`` argument. Common number of\n        dimensions to use for all arrays/tensors.\n        ``theano_function([x, y], [...], dim=2)`` is equivalent to using\n        ``broadcastables={x: (False, False), y: (False, False)}``.\n\n    Returns\n    =======\n    callable\n        A callable object which takes values of ``inputs`` as positional\n        arguments and returns an output array for each of the expressions\n        in ``outputs``. If ``outputs`` is a single expression the function will\n        return a Numpy array, if it is a list of multiple expressions the\n        function will return a list of arrays. See description of the ``squeeze``\n        argument above for the behavior when a single output is passed in a list.\n        The returned object will either be an instance of\n        ``theano.compile.function_module.Function`` or a Python wrapper\n        function around one. In both cases, the returned value will have a\n        ``theano_function`` attribute which points to the return value of\n        ``theano.function``.\n\n    Examples\n    ========\n\n    >>> from sympy.abc import x, y, z\n    >>> from sympy.printing.theanocode import theano_function\n\n    A simple function with one input and one output:\n\n    >>> f1 = theano_function([x], [x**2 - 1], scalar=True)\n    >>> f1(3)\n    8.0\n\n    A function with multiple inputs and one output:\n\n    >>> f2 = theano_function([x, y, z], [(x**z + y**z)**(1/z)], scalar=True)\n    >>> f2(3, 4, 2)\n    5.0\n\n    A function with multiple inputs and multiple outputs:\n\n    >>> f3 = theano_function([x, y], [x**2 + y**2, x**2 - y**2], scalar=True)\n    >>> f3(2, 3)\n    [13.0, -5.0]\n\n    See also\n    ========\n\n    dim_handling\n\n    '
    sympy_deprecation_warning('\n        sympy.printing.theanocode is deprecated. Theano has been renamed to Aesara. Use sympy.printing.aesaracode instead', deprecated_since_version='1.8', active_deprecations_target='theanocode-deprecated')
    if not theano:
        raise ImportError('theano is required for theano_function')
    cache = kwargs.pop('cache', {})
    dtypes = kwargs.pop('dtypes', {})
    broadcastables = dim_handling(inputs, dim=dim, dims=dims, broadcastables=broadcastables)
    code = partial(theano_code, cache=cache, dtypes=dtypes, broadcastables=broadcastables)
    tinputs = list(map(code, inputs))
    toutputs = list(map(code, outputs))
    toutputs = [output if isinstance(output, theano.Variable) else tt.as_tensor_variable(output) for output in toutputs]
    if len(toutputs) == 1:
        toutputs = toutputs[0]
    func = theano.function(tinputs, toutputs, **kwargs)
    is_0d = [len(o.variable.broadcastable) == 0 for o in func.outputs]
    if not scalar or not any(is_0d):
        func.theano_function = func
        return func

    def wrapper(*args):
        if False:
            print('Hello World!')
        out = func(*args)
        if is_sequence(out):
            return [o[()] if is_0d[i] else o for (i, o) in enumerate(out)]
        else:
            return out[()]
    wrapper.__wrapped__ = func
    wrapper.__doc__ = func.__doc__
    wrapper.theano_function = func
    return wrapper